isolated classes and applied OOP principles
