+++ Сompleted_tasks/ClassWork_17_03_22/ClassWork_17_03_22/ClassWork_17_03_22.vcxproj
@@ -141,6 +141,9 @@
   <ItemGroup>
     <ClCompile Include="Main.cpp" />
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Classes.h" />
+  </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
   </ImportGroup>
