+++ Сompleted_tasks/ClassWork_17_03_22/ClassWork_17_03_22/ClassWork_17_03_22.vcxproj.filters
@@ -19,4 +19,9 @@
       <Filter>Исходные файлы</Filter>
     </ClCompile>
   </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="Classes.h">
+      <Filter>Файлы заголовков</Filter>
+    </ClInclude>
+  </ItemGroup>
 </Project>
