+++ Сompleted_tasks/ClassWork_17_03_22/ClassWork_17_03_22/x64/Debug/ClassWork_17_03_22.log
@@ -1,9 +1,9 @@
 ﻿  Main.cpp
-D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Main.cpp(133,32): warning C4244: =: преобразование "double" в "float", возможна потеря данных
-D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Main.cpp(146,26): warning C4244: =: преобразование "double" в "float", возможна потеря данных
-D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Main.cpp(148,26): warning C4244: =: преобразование "double" в "float", возможна потеря данных
-D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Main.cpp(150,26): warning C4244: =: преобразование "double" в "float", возможна потеря данных
-D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Main.cpp(159,16): warning C4244: =: преобразование "double" в "float", возможна потеря данных
-D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Main.cpp(164,46): warning C4244: =: преобразование "int" в "float", возможна потеря данных
-D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Main.cpp(173,119): warning C4244: =: преобразование "double" в "float", возможна потеря данных
+D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Classes.h(209,24): warning C4244: =: преобразование "double" в "float", возможна потеря данных
+D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Classes.h(222,18): warning C4244: =: преобразование "double" в "float", возможна потеря данных
+D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Classes.h(224,18): warning C4244: =: преобразование "double" в "float", возможна потеря данных
+D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Classes.h(226,18): warning C4244: =: преобразование "double" в "float", возможна потеря данных
+D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Classes.h(235,11): warning C4244: =: преобразование "double" в "float", возможна потеря данных
+D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Classes.h(240,44): warning C4244: =: преобразование "int" в "float", возможна потеря данных
+D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\ClassWork_17_03_22\Classes.h(249,114): warning C4244: =: преобразование "double" в "float", возможна потеря данных
   ClassWork_17_03_22.vcxproj -> D:\Project_station_v3\Project station v2\Сompleted_tasks\ClassWork_17_03_22\x64\Debug\ClassWork_17_03_22.exe
